# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App",
        "version": "0.1"
    },
    "paths": {
...
        "/api/cms/categories": {
            "get": {
                "tags": [
                    "CMS Categories"
                ],
                "summary": "Lấy ra danh sách danh muc",
                "operationId": "9e2fa9d524f8b8a8c978a614c85bb17f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per Page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "searchFields[]",
                        "in": "query",
                        "description": "List of fields to search. Example: ['name_booking']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "name_booking"
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Content search. Example: 'Thiên'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Thiên"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter criteria in JSON format. Example: {\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "{\"created_at_RANGE\": [\"2024-01-20\", \"2024-01-28\"]}"
                        }
                    },
                    {
                        "name": "sort[]",
                        "in": "query",
                        "description": "Sort criteria in array format. Use '-' for descending order and '+' for ascending order. Example: ['-created_at']",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "-created_at"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/categories/{id}/": {
            "get": {
                "tags": [
                    "CMS Categories"
                ],
                "summary": "Lấy chi tiet Danh mục",
                "operationId": "197ece1c6c05fe7bd671c6a3693664a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/categories/create": {
            "post": {
                "tags": [
                    "CMS Categories"
                ],
                "summary": "Tạo danh mục mới",
                "operationId": "937a43373aab02762648b951109d33d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/categories/update/{id}": {
            "put": {
                "tags": [
                    "CMS Categories"
                ],
                "summary": "Cập nhật danh mục",
                "description": "Cập nhật thông tin danh mục dựa trên ID",
                "operationId": "e929920399a864acb7555df82a62a591",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ".."
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "do-an"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "danh mục đã được cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "danh mục đã được cập nhật thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cms/categories/delete/{id}": {
            "delete": {
                "tags": [
                    "CMS Categories"
                ],
                "summary": "Xóa danh mục",
                "description": "Xóa danh mục khỏi hệ thống",
                "operationId": "f7129a1459eaf1856e673d83121de2ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "danh mục đã được xoá",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "danh mục đã được xoá thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}